split network into a client and a server
